events: Stabilize image packs
