+++ crates/ruma-events/Cargo.toml
@@ -18,7 +18,6 @@ html = ["dep:ruma-html"]
 markdown = ["dep:pulldown-cmark"]
 unstable-msc1767 = []
 unstable-msc2448 = []
-unstable-msc2545 = []
 unstable-msc2747 = []
 unstable-msc2867 = []
 unstable-msc3061 = []
