+++ crates/ruma-events/src/lib.rs
@@ -163,7 +163,6 @@ pub mod identity_server;
 pub mod ignored_user_list;
 #[cfg(feature = "unstable-msc3552")]
 pub mod image;
-#[cfg(feature = "unstable-msc2545")]
 pub mod image_pack;
 pub mod invite_permission_config;
 pub mod key;
